+++ src/components/CoreConcepts.css
@@ -1,0 +1,43 @@
+#core-concepts {
+    padding: 2rem;
+    border-radius: 6px;
+    background-color: #140524;
+    box-shadow: 0 2px 8px rgba(0, 0, 0, 0.4);
+  }
+  
+  main h2 {
+    text-align: center;
+    font-family: "Roboto Condensed", sans-serif;
+    margin: 0 0 1.5rem 0;
+    color: #a18aba;
+  }
+  
+  #core-concepts ul {
+    list-style: none;
+    padding: 0;
+    margin: 0;
+    display: flex;
+    gap: 2rem;
+    flex-wrap: wrap;
+    justify-content: center;
+  }
+  
+  #core-concepts li {
+    width: 10rem;
+    text-align: center;
+  }
+  
+  #core-concepts li img {
+    height: 4rem;
+    width: 6rem;
+    object-fit: cover;
+  }
+  
+  #core-concepts li h3 {
+    margin: 0.5rem 0;
+  }
+  
+  #core-concepts li p {
+    font-size: 0.9rem;
+  }
+
