+++ src/components/CoreConcepts.jsx
@@ -1,3 +1,5 @@
+import "./CoreConcepts.css"
+
 export default function CoreConcepts ({image, title, description}) {
     return (
         <li>
