+++ src/components/Header.jsx
@@ -1,4 +1,5 @@
 import reactImg from "../assets/react-core-concepts.png"
+import "./Header.css"
 
 function Header(){
