+++ src/components/TabButton.jsx
@@ -1,0 +1,5 @@
+export default function TabButton(props){
+    return(
+        <li><button>{props.children}</button></li>
+    )
+}
